renamed:    .gitignore -> pyramid/.gitignore 	renamed:    README.md -> pyramid/README.md 	new file:   pyramid/__init__.py 	renamed:    brick.py -> pyramid/brick.py 	new file:   pyramid/examples/__init__.py 	renamed:    medium_pyramid.py -> pyramid/examples/medium_pyramid.py 	renamed:    pyramid.py -> pyramid/examples/pyramid.py 	renamed:    small_pyramid.py -> pyramid/examples/small_pyramid.py 	renamed:    triangle.py -> pyramid/examples/triangle.py 	renamed:    graph.py -> pyramid/graph.py 	renamed:    helper_functions.py -> pyramid/helper_functions.py 	renamed:    main.py -> pyramid/main.py 	renamed:    node.py -> pyramid/node.py 	renamed:    preprocessing.py -> pyramid/preprocessing.py 	renamed:    processing.py -> pyramid/processing.py 	new file:   solves/medium_pyramid_solutions.data 	new file:   solves/pyramid_solutions.data 	new file:   solves/small_pyramid_solutions.data 	new file:   solves/triangle_solutions.data 	new file:   testing.py
