+++ pyramid/__init__.py
@@ -1,0 +1,1 @@
+from pyramid.main import get_solutions
