+++ pyramid/examples/medium_pyramid.py
@@ -1,9 +1,9 @@
-from brick import Brick
-from node import Node
-from graph import Graph
-from helper_functions import counter
+from pyramid.brick import Brick
+from pyramid.node import Node
+from pyramid.graph import Graph
+from pyramid.helper_functions import counter
 
-medium_pyramid = Graph("pyramid")
+medium_pyramid = Graph("medium_pyramid")
 
 c = counter()
 # Create all Nodes in a pyramid
