+++ pyramid/helper_functions.py
