+++ pyramid/node.py
